Update webapp.lib.feeds.get_feed
CacheControl is a great library for manging caches based on HTTP headers from the response.
However, some of the feeds, especially the partners feed, don't provide any caching headers,
and so CacheControl won't cache them at all.

requests-cache conversely completely ignores HTTP caching headers and simply lets you set the timeout.

I think in the case of consuming feeds, this is probably a better model. We know how often we want our app making requests, and how fresh the feeds need to be.
We don't want the feed owner to be making that decision for us.

Use the content-type header to decide whether to decode the feed as JSON or as an RSS feed.